final-project/requirements.txt
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-tools==1.43.0